+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - [`rb-scrollspy` component](https://github.com/rockabox/rbx_ui_components/pull/140)
 - [Basic `.SelectControl` styles](https://github.com/rockabox/rbx_ui_components/pull/146)
 - [Add `basics` modifier to `rb-fieldset` component](https://github.com/rockabox/rbx_ui_components/pull/144). Use instead of `campaignBasics`, which will be removed in a future release.
+- [`breakdown and `metrics` modifiers for `rb-fieldset` component](https://github.com/rockabox/rbx_ui_components/pull/155). Intended for use in Reporting UI.
 
 ### Changed
 - [Bumped `karma-jasmine` version to use newer version of `jasmine`](https://github.com/rockabox/rbx_ui_components/pull/154)